requirements.txt
django>=1.6,<1.7
djonet
python-monetdb
astropy
aplpy==0.9.11
numpy
pymongo
matplotlib
pywcs
python-dateutil
psycopg2
GitPython==0.3.2.RC1
pytz

## for caching:
#python-memcached
#pylibmc

# for development
django-debug-toolbar

djangorestframework
markdown       # Markdown support for the REST browsable API.
django-filter  # Filtering support for REST browsable API